mcp-bearer-token/requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.0
asttokens==3.0.0
astunparse==1.6.3
attrs==23.2.0
Authlib==1.6.1
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
cattrs==24.1.3
certifi==2025.6.15
cffi==1.16.0
charset-normalizer==3.4.2
chromadb==1.0.12
click==8.1.7
clipboard==0.0.4
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
comtypes==1.4.8
contourpy==1.3.1
crayons==0.4.0
cryptography==45.0.6
cvzone==1.6.1
cycler==0.12.1
cyclopts==3.22.5
dataclasses-json==0.6.7
debugpy==1.8.11
decorator==5.1.1
diskcache==5.6.3
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.4
django-cors-headers==4.1.0
djangorestframework==3.16.0
dnspython==2.7.0
docstring_parser==0.17.0
docutils==0.22
durationpy==0.10
easyocr==1.7.0
email_validator==2.2.0
exceptiongroup==1.3.0
executing==2.1.0
faiss-cpu==1.11.0
fastapi==0.115.9
fastf1==3.5.3
fastmcp==2.11.2
ffmpeg==1.4
filelock==3.16.1
filterpy==1.4.5
Flask==3.1.1
flatbuffers==24.3.25
fonttools==4.55.2
frozenlist==1.4.1
fsspec==2024.10.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
google-pasta==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.68.1
gTTS==2.5.4
h11==0.16.0
h5py==3.12.1
html5lib==1.1
httpagentparser==1.9.5
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.4
humanfriendly==10.0
idna==3.10
imageio==2.36.1
importlib_metadata==8.7.0
importlib_resources==6.5.2
imutils==0.5.4
infinite_idealist==0.1.2
ipykernel==6.29.5
ipython==8.30.0
ipywidgets==8.1.5
isodate==0.7.2
itsdangerous==2.2.0
jax==0.4.36
jaxlib==0.4.36
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
keras==3.5.0
Keras-Applications==1.0.8
keras-utils==1.0.13
kiwisolver==1.4.7
kubernetes==32.0.1
langchain==0.3.25
langchain-chroma==0.2.4
langchain-community==0.3.24
langchain-core==0.3.64
langchain-text-splitters==0.3.8
langsmith==0.3.45
lazy-object-proxy==1.11.0
lazy_loader==0.4
libclang==18.1.1
licensing==0.41
litellm==1.72.2
llama_cpp_python==0.3.9
lxml==6.0.0
Markdown==3.7
markdown-it-py==3.0.0
markdownify==1.1.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.9.3
matplotlib-inline==0.1.7
mcp==1.12.4
mdurl==0.1.2
mechanize==0.4.9
mediapipe==0.10.18
ml_dtypes==0.5.1
mmh3==5.1.0
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.0.5
mypy_extensions==1.1.0
namex==0.0.8
narwhals==1.41.1
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.2
np-utils==0.6.0
numpy==1.26.4
oauthlib==3.2.2
ollama==0.5.1
onnxruntime==1.22.0
openai==1.93.0
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opencv-contrib-python==4.10.0.84
opencv-python==4.7.0.72
opencv-python-headless==4.10.0.84
opentelemetry-api==1.34.0
opentelemetry-exporter-otlp-proto-common==1.34.0
opentelemetry-exporter-otlp-proto-grpc==1.34.0
opentelemetry-instrumentation==0.55b0
opentelemetry-instrumentation-asgi==0.55b0
opentelemetry-instrumentation-fastapi==0.55b0
opentelemetry-proto==1.34.0
opentelemetry-sdk==1.34.0
opentelemetry-semantic-conventions==0.55b0
opentelemetry-util-http==0.55b0
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.18
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.0.2
parse==1.20.2
parso==0.8.4
pathable==0.4.4
phonenumbers==8.13.28
piexif==1.1.3
Pillow==9.4.0
platformdirs==4.3.6
playsound==1.3.0
posthog==4.4.0
prompt_toolkit==3.0.48
protobuf==4.25.3
psutil==6.1.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycparser==2.21
pycryptodomex==3.20.0
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
pygame==2.3.0
Pygments==2.17.2
PyJWT==2.8.0
PyMuPDF==1.26.0
PyMySQL==1.1.1
PyNaCl==1.5.0
pynput==1.7.1
pyparsing==3.2.0
PyPDF2==3.0.1
pyperclip==1.9.0
PyPika==0.48.9
pypiwin32==223
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PySide6==6.6.1
PySide6-Addons==6.6.1
PySide6-Essentials==6.6.1
PySocks==1.7.1
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pyttsx3==2.98
pytz==2023.3.post1
pywin32==311
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.13.0
readabilipy==0.3.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-cache==1.2.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==14.1.0
rich-rst==1.3.1
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.5.3
scikit-image==0.24.0
scikit-learn==1.6.0
scipy==1.10.1
seaborn==0.13.2
selenium==4.16.0
sentence-transformers==4.1.0
sentencepiece==0.2.0
shapely==2.0.6
shellingham==1.5.4
shiboken6==6.6.1
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soupsieve==2.5
speedtest-cli==2.1.3
SQLAlchemy==2.0.41
sqlparse==0.5.3
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.45.3
stegano==0.11.1
streamlit==1.45.1
sympy==1.13.1
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_keras==0.1
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tiktoken==0.9.0
timple==0.1.8
tokenizers==0.21.1
toml==0.10.2
tor==1.0.0
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
trio==0.24.0
trio-websocket==0.11.1
twilio==9.0.5
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ultralytics==8.0.114
ultralytics-thop==2.0.13
url-normalize==2.2.1
urllib3==2.5.0
utils==1.0.2
uv==0.8.6
uvicorn==0.34.3
watchdog==6.0.0
watchfiles==1.0.5
wcwidth==0.2.13
webbot==0.34
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.1.1
widgetsnbextension==4.0.13
wrapt==1.17.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.22.0
zstandard==0.23.0
